move new code into its own module
