+++ parsl/tests/conftest.py
@@ -1,4 +1,3 @@
-import contextlib
 import importlib.util
 import logging
 import os
@@ -18,10 +17,16 @@ import parsl
 from parsl.dataflow.dflow import DataFlowKernelLoader
 from parsl.tests.utils import get_rundir
 
+# this import doesn't bring in any needed symbols but it does mean
+# that the fixtures declared in logfixtures come into effect.
+# import parsl.tests.logfixtures
+
+pytest_plugins = [
+    "parsl.tests.logfixtures"
+]
+
 logger = logging.getLogger('parsl')
 
-got_bad_log = None
-allow_bad_log = False
 
 def dumpstacks(sig, frame):
     s = ''
@@ -40,43 +45,6 @@ def dumpstacks(sig, frame):
 
 def pytest_sessionstart(session):
     signal.signal(signal.SIGUSR1, dumpstacks)
-
-@pytest.fixture(autouse=True, scope='session')
-def prohibit_severe_logs_session():
-    logger = logging.getLogger()
-    handler = ParslTestLogHandler()
-    logger.addHandler(handler)
-
-
-# this is done after the test has finished rather than directly in the log
-# handler so as to defer the exception until somewhere that pytest will
-# directly see it. If the exception happens inside the log handler, the
-# upwards call stack is not necessarily going to get this properly reported
-# and in some cases causes a hang (!)
-@pytest.fixture(autouse=True)
-def prohibit_severe_logs_test():
-    global got_bad_log
-    yield
-    if got_bad_log is not None:
-        old = got_bad_log
-        got_bad_log = None
-        raise ValueError("Test logged at a severe log level: {}".format(old))
-
- 
-@contextlib.contextmanager 
-def permit_severe_log():
-    global allow_bad_log
-    old = allow_bad_log
-    allow_bad_log = True
-    yield
-    allow_bad_log = old
-
-
-class ParslTestLogHandler(logging.Handler):
-    def emit(self, record):
-        global got_bad_log
-        if record.levelno >= 40 and not allow_bad_log:
-            got_bad_log = "Levelno {}  Levelname {}   Record {}".format(record.levelno, record.levelname, record)
 
 
 def pytest_addoption(parser):
@@ -130,7 +98,6 @@ def pytest_configure(config):
         'markers',
         'issue363: Marks tests that require a shared filesystem for stdout/stderr - see issue #363'
     )
-
 
 
 @pytest.fixture(scope='session')
