+++ parsl/tests/test_bash_apps/test_apptimeout.py
@@ -5,7 +5,8 @@ from parsl.app.app import bash_app
 from parsl.tests.configs.local_threads import config
 from parsl.app.errors import AppTimeout
 
-from parsl.tests.conftest import permit_severe_log
+from parsl.tests.logfixtures import permit_severe_log
+
 
 @bash_app
 def echo_to_file(inputs=[], outputs=[], stderr='std.err', stdout='std.out', walltime=0.5):
