+++ parsl/tests/test_error_handling/test_fail.py
@@ -1,7 +1,7 @@
 import pytest
 
 from parsl.app.app import python_app
-from parsl.tests.conftest import permit_severe_log
+from parsl.tests.logfixtures import permit_severe_log
 
 
 @python_app
